// frontends/epro/src/app/api/model/field-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkflowActionDTO } from './workflow-action-dto';
import { FieldModelDTO } from './field-model-dto';
import { PossibleValueDTO } from './possible-value-dto';
import { WorkflowStatusDTO } from './workflow-status-dto';


/**
 * Fields associated with the dataset
 */
export interface FieldDTO { 
    scopeModelId: string;
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    eventModelId?: string;
    eventPk?: number;
    eventId?: string;
    eventShortname?: string;
    datasetPk: number;
    datasetId: string;
    datasetModelId: string;
    pk: number;
    modelId: string;
    model: FieldModelDTO;
    creationTime: Date;
    lastUpdateTime: Date;
    /**
     * Possible values of the field, if the field type is a multiple
     */
    possibleValues: Array<PossibleValueDTO>;
    value?: string;
    stringValue?: string;
    newContent?: boolean;
    filePk?: number;
    fileName?: string;
    workflowStatus: Array<WorkflowStatusDTO>;
    workflowCreationActions: Array<WorkflowActionDTO>;
    inRemoved: boolean;
}